prototype of subsetter for mongo resource

prototype of subsetter for mongo resource

sbov(rme, gr, map = basicCfieldsMap(), docTypeName = "type", simplify = TRUE)

Arguments

rme
RaggedMongoExpt instance

gr
GRanges instance to subset by

map
list with one element per document type telling what fields are chr, start, stop

docTypeName
character(1) naming column of colData(rme) that has document type

simplify
logical(1) if TRUE, a GRanges is assembled; if FALSE, a RaggedExperiment is returned

Value

if simplify is TRUE (the default), a GRanges is returned with assay components in the mcols; otherwise a RaggedExperiment instance, with metadata component 'docType' set to propagate the input type

Note

This is a placeholder for a proper subsetByOverlaps method. While it would appear sensible to allow subsetting for diverse assay types (e.g., eQTL, FP), in version 1.5.x this is not supported. Each call can operate on only one sample type.

Examples

requireNamespace('mongolite')
if (verifyHasMongoCmd()) {  # for makeColData, which must be able to enumerate collections,
                            # and thus must be able to run system (as opposed to mongolite function) 'mongo' 
 m1 = mongolite::mongo(url=URL_txregLocal(), db='txregnet')
 #cd = makeColData(url=URL_txregLocal(), db='txregnet')
 cd = TxRegInfra2::basicColData.tiny
 rme1 = RaggedMongoExpt(m1, cd[which(cd$type=='FP'),])
 BiocParallel::register(BiocParallel::SerialParam())
 qrng = GRanges('chr1', IRanges(1e6, 1.5e6))
 GenomeInfoDb::genome(qrng) = "hg19"
 ss = sbov(rme1, GRanges('chr1', IRanges(1e6, 1.5e6)), simplify=FALSE)
} 

#> Warning: genome is not set for for query GRanges

#> ..