a GRanges instance with digital genomic footprint metadata returned by `sbov`

a GRanges instance with digital genomic footprint metadata returned by `sbov`

sbov_output_FP

Format

a GRanges instance

Examples

head(sbov_output_FP)

#> GRanges object with 6 ranges and 4 metadata columns:
#>       seqnames            ranges strand |      chr        id      stat
#>          <Rle>         <IRanges>  <Rle> | <factor>  <factor> <numeric>
#>   [1]    chr17 38083253-38083260      * |    chr17 id-454581  0.761905
#>   [2]    chr17 38083323-38083335      * |    chr17 id-454582  0.944444
#>   [3]    chr17 38083535-38083541      * |    chr17 id-454583  0.805556
#>   [4]    chr17 38083577-38083590      * |    chr17 id-454584  0.943787
#>   [5]    chr17 38083593-38083606      * |    chr17 id-454585  0.933258
#>   [6]    chr17 38083719-38083725      * |    chr17 id-454586  0.560784
#>                      origin
#>                 <character>
#>   [1] fLung_DS14724_hg19_FP
#>   [2] fLung_DS14724_hg19_FP
#>   [3] fLung_DS14724_hg19_FP
#>   [4] fLung_DS14724_hg19_FP
#>   [5] fLung_DS14724_hg19_FP
#>   [6] fLung_DS14724_hg19_FP
#>   -------
#>   seqinfo: 1 sequence from hg19 genome